Completed the majority of work on REST API V1. Next steps: testing and fine-tuning code for optimal performance.

+++ app/Http/Controllers/Api/V1/FineController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\FineResource;
+use App\Models\Booking;
 use App\Models\Fine;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
@@ -11,58 +12,40 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class FineController extends Controller
 {
-    /**
-     * The authenticated user
-     *
-     * @var User
-     */
     protected $user;
-
-    /**
-     * The Fine model
-     *
-     * @var Fine
-     */
     protected $model;
 
-    /**
-     * Constructor
-     *
-     * @param Fine $fine The Fine model
-     */
     public function __construct(Fine $fine)
     {
         $this->user = auth()->user();
         $this->model = $fine;
     }
 
-    /**
-     * Display a list of fines for the authenticated user
-     *
-     * @return FineResource fine
-     */
-    public function index(): FineResource
+    public function index()
     {
         try {
-            $fines = $this->model->where('user_id', $this->user->id)->firstOrFail()->load('book');
+            $fines = $this->model->whereUserId($this->user->id)->firstOrFail()->load('book.bookings');
         } catch (ModelNotFoundException $exception) {
             throw new NotFoundHttpException('Do not have any fine');
         }
+
         return FineResource::make($fines);
     }
 
-    /**
-     * Pay the fine for the authenticated user
-     *
-     * @return JsonResponse The success message
-     */
     public function pay(): JsonResponse
     {
         try {
-            $fine = $this->model->where('user_id', $this->user->id)->firstOrFail();
+            $fine = $this->model->whereUserId($this->user->id)->firstOrFail();
         } catch (ModelNotFoundException $exception) {
             throw new NotFoundHttpException('Do not have any fine');
         }
+
+        $bookingWasReturned = Booking::withTrashed()->whereUserId($this->user->id)->where('status', 'active')->exists();
+
+        if ($bookingWasReturned) {
+            throw new NotFoundHttpException('You have not returned the book yet, please return it first');
+        }
+
         $fine->delete();
         return response()->json(['message' => 'Fine paid successfully']);
     }
